Safari-609.1.11.1/LayoutTests/fast/animation/request-animation-frame-throttle-subframe-display-none.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that requestAnimationFrame is not throttled in subframes that are display:none");
window.jsTestIsAsync = true;

function runTest()
{
    testFrame = document.getElementById("testFrame");

    // Force layout.
    document.body.offsetTop;

    shouldBeFalse("testFrame.contentWindow.internals.isRequestAnimationFrameThrottled()");
    finishJSTest();
}
</script>
<iframe id="testFrame" style="position: absolute; top: 800px; display: none;" src="resources/requestAnimationFrame-frame.html" onload="runTest()"></iframe>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
